requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-silk==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-spectacular==0.21.1
gprof2dot==2021.2.21
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.2
libcst==0.3.23
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
